move client to lib to allow further testing
